+++ code/CMakeLists.txt
@@ -32,6 +32,22 @@ set(UISources
 )
 source_group("Source/UISources" FILES ${UISources})
 
+set(SynthSources
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/Grain.h
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/Grain.cpp
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/Scheduler.h
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/Scheduler.cpp
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/SequenceStrategy.h
+    ${CMAKE_SOURCE_DIR}/dsp/synthesis/granular/SequenceStrategy.cpp
+)
+source_group("Source/SynthSources" FILES ${UISources})
+
+set(EnvelopSources
+    ${CMAKE_SOURCE_DIR}/dsp/envelopes/Envelope.h
+    ${CMAKE_SOURCE_DIR}/dsp/envelopes/Envelope.cpp
+)
+source_group("Source/EnvelopeSources" FILES ${UISources})
+
 set(DspSources
     ${CMAKE_SOURCE_DIR}/dsp/OfflineStretcher.h
     ${CMAKE_SOURCE_DIR}/dsp/OfflineStretcher.cpp
