+++ code/dsp/synthesis/granular/SequenceStrategy.cpp
@@ -1,0 +1,27 @@
+#include "SequenceStrategy.h"
+
+SequenceStrategy::SequenceStrategy()
+{
+    
+}
+
+void SequenceStrategy::setGrainDuration(size_t duration)
+{
+    mDuration.store(duration);
+}
+
+void SequenceStrategy::setGrainDensity(double grainDensity)
+{
+    mGrainsPerUnitTime.store(grainDensity);
+}
+
+size_t SequenceStrategy::nextDuration()
+{
+    return mDuration.load();
+}
+
+float SequenceStrategy::nextInteronset()
+{
+    return -std::log(mRandom.nextFloat()) / mGrainsPerUnitTime.load();
+}
+
